move lds.py and attributions/
